Change repo_upload_file field type to hidden instead of text to improve browser compat
Firefox 3.6 would bug out of the upload code without sending the file to the server, because it was silently failing to set the value on this field with type="text".  Changing to type="hidden" gets around this and allows it to upload cleanly, and doesn't seem to break other browsers.